+++ Metadec/Assets/Plugins/WebView/Scripts/WebView.cs
@@ -1,11 +1,18 @@
 using System;
-using Newtonsoft.Json;
 using UnityEngine;
+using Newtonsoft.Json;
+using System.Threading.Tasks;
+
+#if UNITY_ANDROID
+    using UnityEngine.Android;
+#endif
 
 namespace ReadyPlayerMe
 {
     public class WebView : MonoBehaviour
     {
+        private static UserAgent userAgent = null;
+        
         private const string DATA_URL_FIELD_NAME = "url";
         private const string AVATAR_EXPORT_EVENT_NAME = "v1.avatar.exported";
 
@@ -42,17 +49,14 @@ namespace ReadyPlayerMe
                     messagePanel.SetMessage(MessagePanel.MessageType.NotSupported);
                     messagePanel.SetVisible(true);
                 #else
-                    if (webViewObject == null)
-                    {
-                        messagePanel.SetMessage(MessagePanel.MessageType.Loading);
-                        messagePanel.SetVisible(true);
+                    messagePanel.SetMessage(MessagePanel.MessageType.Loading);
+                    messagePanel.SetVisible(true);
 
-                        #if UNITY_ANDROID
-                            webViewObject = gameObject.AddComponent<AndroidWebViewWindow>();
-                        #elif UNITY_IOS
-                            webViewObject = gameObject.AddComponent<IOSWebViewWindow>();
-                        #endif
-                    }
+                    #if UNITY_ANDROID
+                        webViewObject = gameObject.AddComponent<AndroidWebViewWindow>();
+                    #elif UNITY_IOS
+                        webViewObject = gameObject.AddComponent<IOSWebViewWindow>();
+                    #endif
 
                     webViewObject.OnLoaded = OnLoaded;
                     webViewObject.OnJS = OnWebMessageReceived;
@@ -70,6 +74,89 @@ namespace ReadyPlayerMe
             SetScreenPadding(left, top, right, bottom);
         }
 
+        /// User Agent of the system WebView of the device, that will be used for displaying RPM website.
+        private async Task<UserAgent> GetUserAgent()
+        {
+            var timeout = 5f;
+            var resolved = false;
+            UserAgent webViewUserAgent = new UserAgent();
+            
+            #if UNITY_EDITOR || !(UNITY_ANDROID || UNITY_IOS)
+                messagePanel.SetMessage(MessagePanel.MessageType.NotSupported);
+                messagePanel.SetVisible(true);
+                return webViewUserAgent;
+            #else
+                #if UNITY_ANDROID
+                    webViewObject = gameObject.AddComponent<AndroidWebViewWindow>();
+                #elif UNITY_IOS
+                    webViewObject = gameObject.AddComponent<IOSWebViewWindow>();
+                #endif
+            #endif
+
+            // called when page sends UserAgent string
+            webViewObject.OnJS = (message) =>
+            {
+                webViewUserAgent = new UserAgent(message);
+                Destroy(webViewObject);
+                resolved = true;
+            };
+
+            // called when blank page is loaded 
+            webViewObject.OnLoaded = (message) =>
+            {
+                webViewObject.EvaluateJS(@"
+                    if (window && window.webkit && window.webkit.messageHandlers && window.webkit.messageHandlers.unityControl) {
+                        window.Unity = { call: function(msg) { window.webkit.messageHandlers.unityControl.postMessage(msg); } }
+                    }
+                    else {
+                        window.Unity = { call: function(msg) { window.location = 'unity:' + msg; } }
+                    }
+                    
+                    Unity.call(window.navigator.userAgent);
+                ");
+            };
+            
+            WebViewOptions options = new WebViewOptions();
+            webViewObject.Init(options);
+            
+            if (!webViewObject.IsWebViewAvailable())
+            {
+                Destroy(webViewObject);
+                return webViewUserAgent;
+            }
+            
+            webViewObject.LoadURL("");
+
+            while (!resolved)
+            {
+                if (timeout <= 0)
+                {
+                    Debug.Log("WebView check timed out.");
+                    break;
+                }
+                
+                timeout -= Time.deltaTime;
+                await Task.Yield();
+            }
+
+            return webViewUserAgent;
+        }
+
+        /// <summary>
+        ///     Check if the WebView that will be used for displaying RPM website is up to date for handling 3D graphics.
+        ///     If it returns false, it is not advised to load WebView since the browser performance will be poor.
+        ///     Instead, you can warn the user for them to update their system WebView applications.
+        /// </summary>
+        public async Task<bool> IsWebViewUpToDate()
+        {
+            if (userAgent == null)
+            {
+                userAgent = await GetUserAgent();
+            }
+            
+            return userAgent.IsWebViewUpToDate();
+        }
+        
         /// <summary>
         ///     Set WebView screen padding in pixels.
         /// </summary>
@@ -183,10 +270,29 @@ namespace ReadyPlayerMe
                 );
 
                 window.removeEventListener('message', subscribe);
-                window.addEventListener('message', subscribe)
+                window.addEventListener('message', subscribe);
             ");
 
+            AskForCameraPermission();
+
             Loaded = true;
+        }
+
+        private void AskForCameraPermission()
+        {
+            #if UNITY_ANDROID
+                if (!Permission.HasUserAuthorizedPermission(Permission.Camera))
+                {
+                    Permission.RequestUserPermission(Permission.Camera);
+                }
+            #endif
+                
+            #if UNITY_IOS
+                if (!Application.HasUserAuthorization(UserAuthorization.WebCam))
+                {
+                    Application.RequestUserAuthorization(UserAuthorization.WebCam);
+                }
+            #endif
         }
 
         private void OnDrawGizmos()