+++ Metadec/Assets/Scripts/InstanciarJugadpr.cs
@@ -1,5 +1,6 @@
 using Photon.Pun;
 using Photon.Realtime;
+using ReadyPlayerMe;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -11,17 +12,22 @@ public class InstanciarJugadpr : MonoBehaviourPunCallbacks
     public GameObject carga;
     public GameObject avatar;
     public GameObject playerAvater;
-    public GameObject mainCamera;
-    public GameObject playerFollow;
+    //public GameObject prefMainCamera;
+    //public GameObject prefPlayerFollow;
     public bool cargandoJugadores = true;
     public int nJugadores;
     public int nJugadoresEnEscena = 0;
+
+    public GameObject mainCamera;
+    public GameObject playerFollow;
 
     public List<string> avatarUrls = new List<string>();
     public List<string> namesPlayersInstantiate = new List<string>();
 
     void Start()
     {
+        QualitySettings.vSyncCount = 0;
+        Application.targetFrameRate = -1;
         carga.SetActive(true);
         if (!PhotonNetwork.IsConnected)
         {
@@ -30,9 +36,10 @@ public class InstanciarJugadpr : MonoBehaviourPunCallbacks
             return;
         }
 
-        Instantiate(mainCamera, Vector2.zero, Quaternion.identity);
-        Instantiate(playerFollow, Vector2.zero, Quaternion.identity);
+        //Instantiate(prefMainCamera, Vector2.zero, Quaternion.identity);
+        //Instantiate(prefPlayerFollow, Vector2.zero, Quaternion.identity);
         avatar = PhotonNetwork.Instantiate(playerAvater.name, new Vector3(Random.Range(30f, 58.52f), 2.3f, Random.Range(43.8f, 50f)), Quaternion.identity, 0) as GameObject;
+        //avatar = PhotonNetwork.Instantiate(playerAvater.name, new Vector3(441.399994f, 0.0583543777f, 246.270004f), Quaternion.identity, 0) as GameObject;
         avatar.GetComponent<EnEscenario>().control = this.gameObject;
         nJugadores = (int)PhotonNetwork.PlayerList.Length;
     }
@@ -49,7 +56,7 @@ public class InstanciarJugadpr : MonoBehaviourPunCallbacks
                 {
                     avatarUrls.Add(player.CustomProperties["avatar"].ToString());
                 }
-                cargarAvatars.GetComponent<ReadyPlayerMe.RuntimeExampleMultiple>().avatarUrls.AddRange(avatarUrls);
+                cargarAvatars.GetComponent<RuntimeExampleMultiple>().avatarUrls.AddRange(avatarUrls);
                 //cargarAvatars.GetComponent<ReadyPlayerMe.RuntimeExampleMultiple>().cargarAvatarsEscena();
                 cargandoJugadores = false;
             }